+++ lib/http_service/builders/only.rb
@@ -1,0 +1,25 @@
+# frozen_string_literal: true
+
+module HTTPService
+  module Only
+    module Body
+      include Requester
+
+      private
+
+      def after_request
+        super.body
+      end
+    end
+
+    module Headers
+      include Requester
+
+      private
+
+      def after_request
+        super.headers
+      end
+    end
+  end
+end