+++ lib/http_service/builders/raise_error.rb
@@ -1,0 +1,14 @@
+# frozen_string_literal: true
+
+module HTTPService
+  module RaiseError
+    include ConfigurableClient
+
+    private
+
+    def configure_client(faraday)
+      super
+      faraday.response(:raise_error)
+    end
+  end
+end